requirements.txt
pandas
numpy
plotly
streamlit
folium
streamlit-folium
pillow
haversine